Simplified configuration for starter. only depends on :starter key.
Removed multi-methods 'start' and 'initialize' from starter.
the entry point for starting a service now is a simple function, and
supplied as a symbol or string of a function with :starter key in our
configuration file.

crow.starter simplly load the function supplied with :starter key and
execute it.

Samely, :object-marshaller can be defined as a symbol or string of
a constructor function. crow.starter will load and execute the fn
if the value of :object-marshaller is a symbol or string.